requirements.txt
requests>=2.31.0
yfinance>=0.2.28
textblob>=0.17.1
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
